requirements.txt
torch
torchvision
gymnasium[box2d]
opencv-python
imageio
matplotlib
tqdm
tensorboard
jupyter
ipywidgets